+++ NativeVsfClone/unity/Packages/com.vsfclone.xav2/Editor/Xav2ExportOptions.cs
@@ -20,6 +20,8 @@ namespace VsfClone.Xav2.Editor
         public bool FailOnMissingShader = true;
         public List<string> StrictShaderSet = new List<string>()
         {
+            "Standard",
+            "MToon",
             "lilToon",
             "Poiyomi"
         };